tests: move setup/helper code to separate file
This way you can do `include("test/setup.jl")` and then run any of
the testset groups from the main file since they are independent.
